feat(router)!: add page routing and restructure App component for improved navigation
This is a significant change for the entire website, applying React Router with react-dom to enhance navigation. The new sections include:

Gear (showcasing my home hardware like phone, home server, desktop and and laptop.
Project (displaying my projects)
Blog (featuring articles related to this website)
About (a section dedicated to UmmIt with an in-depth overview)

With list and other changed:

- Implemented React Router for seamless navigation between the new sections.
- Replaced the hero section in App.jsx with Routes for dynamic content rendering.
- Updated Navbar to utilize Link from react-router-dom for client-side navigation and improved mobile/desktop styles.
- Added a heart icon to the Footer for visual appeal.
- Wrapped the App component in BrowserRouter in main.jsx for routing support.